+++ changelog.md
@@ -32,6 +32,7 @@
   - Added tests.
 - Removed the test helper class `CountReadableStream` and refactored the function `ignoreStream` tests.
 - Test that `GraphQLUpload` is a GraphQL scalar.
+- In tests, use `getDefaultHighWaterMark` from `node:stream` to avoid hardcoded assumptions about chunk sizes.
 - Fixed changelog entry typos for v4.0.0-alpha.2, 6.0.0-alpha.1, and v17.0.0.
 
 ## 17.0.0